Adding capability such that climate can be run with any metallicity and C/O
This is mainly based on interpolating the chemistry in between grid points from the 1460 tables. So, if the user runs exact values of metallicity and C/O for which sonora tables already are present, then this will be based on exact calculation. Otherwise, chemistry will be interpolated. Interpolation is needed mainly for alkalis which are not in chemical networks.